refactor: simplify auth context folder structure
