--- src/domain/contexts/authContext/components/index.ts
@@ -1,1 +1,0 @@
-export * from './AuthContextProvider';
