+++ src/domain/contexts/authContext/useAuthContext.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import AuthContext from '../AuthContext';
+import AuthContext from './AuthContext';
 
 export const useAuthContext = () => {
   const value = useContext(AuthContext);
